+++ jobboard-core/src/main/java/com/advanceweb/afc/jb/jobseeker/service/JobSeekerSubscriptions.java
@@ -22,13 +22,13 @@ public interface JobSeekerSubscriptions {
 	 * @return
 	 */
 
-	public boolean saveJobSeekerSubscription(List<JobSeekerSubscriptionsDTO> listSubsDTO, long userId);
+	public boolean saveJobSeekerSubscription(List<JobSeekerSubscriptionsDTO> listSubsDTO, int userId);
 	
 	/**
 	 * To get current subscriptions of the user
 	 * @param userId
 	 * @return
 	 */
-	public List<JobSeekerSubscriptionsDTO> getCurrentSubscriptions(long userId);
+	public List<JobSeekerSubscriptionsDTO> getCurrentSubscriptions(int userId);
 
 }
